demo-maven-plugin/maven-generate-file-plugin/src/test/java/com/example/GenerateFileMojoTest.java
package com.example;

import org.apache.maven.plugin.testing.AbstractMojoTestCase;

import java.io.File;

public class GenerateFileMojoTest extends AbstractMojoTestCase {

    protected void setUp() throws Exception {
        super.setUp();
    }

    protected void tearDown() throws Exception {
        super.tearDown();
    }

    public void testSomething() throws Exception {
        File pom = getTestFile("src/test/resources/pom.xml");
        assertNotNull(pom);
        assertTrue(pom.exists());
        GenerateFileMojo myMojo = (GenerateFileMojo) lookupMojo("generate-file", pom);
        assertNotNull(myMojo);
        myMojo.execute();
    }
}